pr feedback. add larger limit to api token GET
